Remove external specs file
